requirements.txt
Django==3.0
django-bootstrap3==14.2.0
django-bootstrap4==2.3.1
django-braces==1.14.0
django-crispy-forms==1.9.2
django-multiselectfield==0.1.12
django-weekday-field==1.1.0
Flask==1.1.2
Pillow==7.2.0
whitenoise==5.2.0
asgiref==3.2.10
python-dateutil==2.8.1
pytz==2020.1
sqlparse==0.3.1
urllib3==1.25.10